refactor: make util for copy to clipboard and move everything under src
