+++ src/backend/schemas/memory.schema.js
