+++ tailwind.config.js
@@ -3,7 +3,7 @@ const { heroui } = require('@heroui/theme')
 export default {
   content: [
     './pages/**/*.{js,ts,jsx,tsx}',
-    './components/**/*.{js,ts,jsx,tsx}',
+    './src/**/*.{js,ts,jsx,tsx}',
     './node_modules/@heroui/theme/dist/components/(button|card|image|select|ripple|spinner|form|listbox|divider|popover|scroll-shadow).js',
   ],
   theme: {
